Move string settings to StringSettings
